Started refactoring of tests using tox.
