--- password_policies/tests/__init__.py
@@ -1,5 +1,0 @@
-from password_policies.tests.forms import *
-from password_policies.tests.middleware import *
-from password_policies.tests.models import *
-from password_policies.tests.utils import *
-from password_policies.tests.views import *
