+++ tox.ini
@@ -1,0 +1,12 @@
+[tox]
+envlist =
+    py27-{17,18,19},
+    py33-{17,18},
+    py34-{17,18,19},
+    py35-{18,19}
+[testenv]
+deps =
+    17: Django >= 1.7, < 1.8
+    18: Django >= 1.8, < 1.9
+    19: Django >= 1.9, < 1.10
+commands = python setup.py test
